Classes

+++ src/main/scala/ayai/components/Room.scala
@@ -3,9 +3,9 @@ package ayai.components
 /** Crane Imports **/
 import crane.Component
 
-class Room(val id: Int) extends Component {
+class Room(val id: Long) extends Component {
   //these are used to help respawn or switch rooms
-  // val prevId : Int = _
+  // val prevId : Long = _
   // val prevEndX : Int = _
   // val prevEndY : Int = _
 }
