+++ src/main/scala/ayai/systems/MovementSystem.scala
@@ -12,7 +12,8 @@ import scala.collection.mutable.HashMap
 import org.slf4j.{Logger, LoggerFactory}
 import java.util.Map
 
-class MovementSystem(roomHash : HashMap[Int, Entity]) extends EntityProcessingSystem(include=List(classOf[Position], classOf[Velocity],classOf[Room], classOf[Character]), exclude=List(classOf[Transport], classOf[Respawn])) {	
+
+class MovementSystem(roomHash : HashMap[Long, Entity]) extends EntityProcessingSystem(include=List(classOf[Position], classOf[Velocity],classOf[Room], classOf[Character]), exclude=List(classOf[Transport], classOf[Respawn])) {	
   private val log = LoggerFactory.getLogger(getClass)
 	  //this will only move characters who have received a movement key and the current component is still set to True
 	  override def processEntity(e: Entity, delta : Int) {
